refactor: moving buttons into its own file
